рефакторинг + фиксы

+++ Backend/SteamTradeBot.Backend/BusinessLogicLayer/Solutions/CancelMarketSolution.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Serilog;
 using SteamTradeBot.Backend.DataAccessLayer;
 using SteamTradeBot.Backend.Models.Abstractions;
@@ -6,7 +7,7 @@ using SteamTradeBot.Backend.Models.ItemModel;
 
 namespace SteamTradeBot.Backend.BusinessLogicLayer.Solutions;
 
-public class CancelMarketSolution : MarketSolution
+public sealed class CancelMarketSolution : MarketSolution
 {
     public CancelMarketSolution(
         ISteamApi api, 
@@ -17,15 +18,18 @@ public class CancelMarketSolution : MarketSolution
 
     public override void Perform(ItemPage itemPage)
     {
-        throw new System.NotImplementedException();
+        throw new NotImplementedException();
     }
 
     public override async Task PerformAsync(ItemPage itemPage)
     {
+        Log.Information("Cancelling buy order ({0})...", itemPage.EngItemName);
         var order = await MarketDb.GetBuyOrderAsync(itemPage.EngItemName, ConfigurationManager.ApiKey);
+        if (order is null)
+            throw new Exception("Can't load stored buy order from database.");
         await SteamApi.CancelBuyOrderAsync(order.ItemUrl);
-        Log.Information("Cancel buy order {0} (Price: {1}, Quantity: {2})", order.EngItemName, order.Price, order.Quantity);
         await MarketDb.RemoveBuyOrderAsync(order);
         await StateManager.OnItemCancellingAsync(order);
+        Log.Information("Buy order {0} (Price: {1}, Quantity: {2}) has been canceled.", order.EngItemName, order.BuyPrice, order.Quantity);
     }
 }
